Remove unused modules, restructure code.
