+++ src/characters/character_examples.py
